Add code cloned from liblfsxx
Create copies of several files with utility functions I implemented in
recent months inside liblfsxx as they are needed by the code that is to
be added to nosync. Adjust namespace name and remove macro wrapper for
"warn_unused_result" attribute where it's used.

In the future some parts of this code will be removed as some variants
of functions are not really needed in nosync. Also, some functions will
be removed after eliminating their usage from nosync.